refactor: moved test case
